Updated depreciated/index.html and resources/index.html
Updated the "index.html" file of the "depreciated" directory with a whole new theme featuring a new color for the header and a background image of a campfire (flame_img(DSC_0284).JPG). Updated the name of the "index.html" file for the "resources" directory to: "WISPNET – Resources" instead of the incorrect "WISPNET – Depreciated" which was accidentally carried over by a copy.